InventoryManagement: Fix bug related to search hotkey
The search hotkey event handler called e.preventDefault() before the
guard to verify which key was pressed on keyboard. This adds the problem
that when this event is bound, all other keys may have their
functionality prevented. Therefore, move e.preventDefault() in to the
guard where the hotkey is checked.