Inventory: Correctly subtract inventory amount
The .remove() function did not decrease the total count of the
inventory. To ensure this, we always remove items when added to
inventory, and to make the .remove() function always decrease the
amount. In this way, we correctly decrease the stack amount of the item
upon removal and correctly handle the case in which a item with a 0
stack is first removed and not added back.