IngredientStack: Move stack clamping to the IngredientStack class
In BOTW, material items may stack up to 999 times. Since IngredientStack
is supposed to represent a item + stack, it makes sense to move clamping
logic to this class. The ClickableIngredient should not be responsible
for maintaining this logic.

Lastly, ClickableIngredient should default to zero on NaN input than to
999. This problem was caused when hitting backspace on the number too
much. This does result in a leading zero, when putting in a new number,
but this is OK.