Introduce the concept of an Item
It's becoming important that we're able to represent recipes and
ingredients as items. We therefore introduce an item interface. This
abstracts away the common functionality of recipe from an ingredient
but also allow for stacking of both items.

This is not helpful _now_ but we want to be ready for when we allow
users to bulk-craft recipes.

Unfortunately, this system will not take into account if your inventory
will be "full" or not, but nevertheless, users should be able to want to
create as many "stacks" of plates as they want despite the fact they
don't stack in game and take up additional inventory space.