Huge refactor
When attempting to add recipes for search, there became an issue with
coming up with a way to represent recipes that would allow for multiple
types of ingredients in different combinations. This new design
addresses that by making the recipes matching based on a categorical
system and new modeling. To simplify, recipes are actually not items in
the grand scheme of things, they're instead a collection of rules that
result in an item, a consumable. Recipes need ingredients, but it's more
correct to classify these ingredients as materials.

With this in mind, we make materials and consumables implement item. We
make recipes define their properties with categories on the items. We
introduce a special ItemCategory type to represent these categories.

In this way we improve our design by making a easy string-based system
to build indexes off of to create our combinations of consumables via
rules defined by recipes.

How this will look in the end is yet to be determined, but I'm thinking
that we can use the loose-rules of recipes to determine _what_ can be
made, but then once we have the recipes we can create, we then expand
the result set to include the resulting consumables that will have the
appropriate effects based off the items selected. These true items will
be what is in our index. I'm thinking we'd have dynamically building
indexes. Ideally this doesn't take _too much time_ from a processing
standpoint.