Recipe: Implement 2/3rds of recipes
In addition to adding in 2/3rds of the recipes, also add updated logic
for figuring out what things users can make. There's still a few edge
cases that are not handled, and therefore will be addressed at
some point. For now, the majority of the recipes seem to be calculated
correctly.