Recipe: Introduce inventory make up of a recipe
At some point to make search criteria viable, we need to know what
ingredients triggered a recipe to be created. In this way, we can expand
our result set to build dishes with their varying effects.

A bug was however introduced such that we're treating categories as
materials when this shouldn't be the case. Therefore, in a future
commit, the problem will be resolved by making the matching a little bit
more intelligent when it comes to matching ingredients to categories.