Project: Theme the app to more closely to in-game inventory
This initial style of the app isn't a 1-1 with the inventory in game,
but its closer now than it was before. For a speed increase, we use an
already pre-blurred & compressed background image for load than to apply
CSS filter: blur() apparently, blur() is an expensive operation for the
browser.