Recipe: Differentiate between craftable recipe and non-craftable recipe
Currently we're not keeping track of how many of a recipe we can craft.
Also, the current design was meant as a temporary stepping stone before
arriving here at our final recipe class. The makeup() function will now
return a recipe with only certain properties set if that recipe could be
crafted. Later, this information will be used to construct our expanded
result set w/ effects to search on.