Massively update inventory to recipe matching
This is a complete overhaul of the matching algorithms. There are two
problems to overcome:

  1. Create every combination of a recipe given inventory
  2. Perform matching

We overcome problem 1 by first creating a separate script to be ran and
prior to running the code, the compile-recipes script. This script is
necessary because calculating every possible combination of a recipe at
runtime is too prohibitive. Therefore, we pre-generate all possible
combinations into a recipes.ts database file for later access.

The interesting trick to problem 1 was to run a pre-pass on the recipes
ensuring that all the categories are fully filled out. The next pass
takes the logic tree, and constructs a "forest". Because of this, we're
then able to trivially pick out the required items & then could perform
the algorithm that finally creates every possible combination to a
recipe.

For problem 2, various indexes are built at "first thing" during runtime
to perform set-based mapping of inventory to recipes.

Overall, problem 2 is still lacking in the performance area, and there's
still a problem with problem 1 such that recipes with more than one of
the same ingredient need to be bundled together.

This commit is being made now because this is a huge huge overhaul, and
remaining updates should be their own commits. Since the "build" is not
broken at this commit, ship it.