Make performance improvements
Commit d2b57d7 ("Massively update inventory to recipe matching")
introduced a tool to generate all possible combinations of recipes
based on their logic. One of the lingering problems was duplication of
recipe items & duplication of recipes, because the ingredients could be
in a different order. Fix this by tracking the unique recipes by
ordering the ingredients in their id's and then also double up on
ingredients already existing in the recipe combinations.

Lastly, remove some console.log()'s that are no longer needed.